+++ src/services/database/repository.ts
@@ -1,0 +1,8 @@
+import StoredPlayerSchema from '../../base/schemas/StoredPlayerSchema';
+
+export const getStoredPlayersByDiscordIds = async (ids: string[]) => {
+  const players = await StoredPlayerSchema.find({
+    discordId: { $in: ids },
+  }).lean();
+  return players;
+};