+++ CHANGELOG.md
@@ -1,15 +1,16 @@
-## [1.8.10](https://github.com/Deepwerks/discord-bot/compare/v1.8.5...v1.8.10) (2025-06-21)
+## [1.8.10](https://github.com/Deepwerks/discord-bot/compare/v1.8.5...v1.8.10) (2025-06-27)
 
 
 ### Bug Fixes
 
-* default lobby name ([dcc6e11](https://github.com/Deepwerks/discord-bot/commit/dcc6e1138b75f2f534cd9e79282863e9678805f2))
+* default lobby name ([#227](https://github.com/Deepwerks/discord-bot/issues/227)) ([c1529bc](https://github.com/Deepwerks/discord-bot/commit/c1529bce05bf0e91fe594dc607662d9711cccc83))
 * fixed bearer token read ([#218](https://github.com/Deepwerks/discord-bot/issues/218)) ([ec97bc5](https://github.com/Deepwerks/discord-bot/commit/ec97bc592d0aa362db902e7eaa140dd2f67cf2c5))
 * updated schema validator on create-lobby request ([#222](https://github.com/Deepwerks/discord-bot/issues/222)) ([370cee1](https://github.com/Deepwerks/discord-bot/commit/370cee15fb3dcbb0472bca3029d44c6cc60c07ed))
 
 
 ### Features
 
+* added balanced shuffle option ([b48266d](https://github.com/Deepwerks/discord-bot/commit/b48266d087719b0da63a9a9cd31d1517ded52ad8))
 * reintroduced metrics ([#216](https://github.com/Deepwerks/discord-bot/issues/216)) ([8f9f2fb](https://github.com/Deepwerks/discord-bot/commit/8f9f2fbdb68429a596cb48095159b8e4dc32a694))